Adding capability to configure private Route53 domains
At times a user will want to configure private Route53 domains.
This introduces the capbility to set a flag and have terraform
create the private Route53 DNS entries.

This feature is useful to configure such things as split DNS and
also allow for private DNS entries for such things as nexus and
chartmuseum.